+++ src/net/peacefulcraft/sco/gamehandle/player/SCOPlayer.java
@@ -7,7 +7,6 @@ import org.bukkit.GameMode;
 import org.bukkit.entity.Player;
 
 import net.md_5.bungee.api.ChatColor;
-import net.peacefulcraft.sco.gamehandle.storage.SCOPlayerDataManager;
 import net.peacefulcraft.sco.swordskills.skills.SwordSkill;
 
 public class SCOPlayer
